requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
pymysql
pinecone-client
openai
langchain
pandas

tiktoken